+++ src/test/java/school/redrover/DashboardTest.java
@@ -2,7 +2,11 @@ package school.redrover;
 import org.openqa.selenium.By;
 import org.testng.Assert;
 import org.testng.annotations.Test;
+import school.redrover.model.DashboardPage;
 import school.redrover.runner.BaseTest;
+
+import java.util.Arrays;
+import java.util.List;
 
 public class DashboardTest extends BaseTest {
 
@@ -12,5 +16,20 @@ public class DashboardTest extends BaseTest {
                    .click();
         String idText = getDriver().findElement(By.xpath("//*[contains(text(), 'ID')]")).getText();
         Assert.assertTrue(idText.contains("Jenkins User ID"));
+    }
+
+    @Test
+    public void testDashboardMenu() {
+        final List<String> expectedDashboardMenu = Arrays.asList(
+                "New Item",
+                "People",
+                "Build History",
+                "Manage Jenkins",
+                "My Views");
+
+        List<String> actualDashboardMenu = new DashboardPage(getDriver())
+                .getDashboardMenuList();
+
+        Assert.assertEquals(actualDashboardMenu, expectedDashboardMenu );
     }
 }
